+++ tests/testthat/test-layer_spatraster.R
@@ -8,7 +8,7 @@ test_that("Test layers", {
   expect_error(layer_spatraster(x))
 
   skip_if_not_installed("ggplot2")
-  skip_if_not_installed("ggspatial")
+  skip_if_not_installed("tidyterra")
   skip_if_not_installed("raster")
   skip_if_not_installed("slippymath")
   skip_if_not_installed("terra")
@@ -20,21 +20,5 @@ test_that("Test layers", {
   tile <- esp_getTiles(x)
 
 
-  expect_silent(layer_spatraster(tile))
-
-  g <- ggplot2::ggplot() +
-    layer_spatraster(tile)
-
-
-  expect_s3_class(g, "ggplot")
-  t <- tempfile(fileext = ".png")
-
-  # Run only locally
-  skip_on_ci()
-
-  suppressWarnings(ggplot2::ggsave(t, g,
-    width = 3, height = 3,
-    dpi = 90
-  ))
-  expect_snapshot_file(t, "layer_spatraster.png")
+  expect_error(layer_spatraster(tile))
 })